requirements.txt
asgiref==3.3.1
certifi==2020.11.8
chardet==3.0.4
Django==3.1.3
django-ckeditor==6.0.0
django-ckeditor-5==0.0.12
django-crispy-forms==1.10.0
django-js-asset==1.2.2
djangorestframework==3.12.2
fcm-django==0.3.4
gunicorn==20.0.4
idna==2.10
Jinja2==2.11.2
jinja2-django-version==1.1.2
MarkupSafe==1.1.1
Pillow==8.0.1
pkg-resources==0.0.0
pyfcm==1.4.7
pytz==2020.4
requests==2.25.0
sqlparse==0.4.1
urllib3==1.26.2
whitenoise==5.2.0
